+++ aws/doc/gen_paramsdoc.go
@@ -298,6 +298,7 @@ var generatedParamsDoc = map[string]map[string]string{
 	"deletebucket": {
 		"name": "",
 	},
+	"deletecontainer": {},
 	"deletecontainercluster": {
 		"id": "The short name or full Amazon Resource Name (ARN) of the cluster to delete",
 	},
@@ -446,17 +447,33 @@ var generatedParamsDoc = map[string]map[string]string{
 	"startalarm": {
 		"names": "The names of the alarms",
 	},
-	"startcontainerservice": {},
+	"startcontainerservice": {
+		"cluster":         "The short name or full Amazon Resource Name (ARN) of the cluster on which to run your service",
+		"deployment-name": "The name of your service",
+		"desired-count":   "The number of instantiations of the specified task definition to place and keep running on your cluster",
+		"name":            "The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to run in your service",
+		"role":            "The name or full Amazon Resource Name (ARN) of the IAM role that allows Amazon ECS to make calls to your load balancer on your behalf",
+	},
 	"startinstance": {
 		"id": "One or more instance IDs",
 	},
 	"stopalarm": {
 		"names": "The names of the alarms",
 	},
+	"stopcontainerservice": {
+		"cluster":         "The name of the cluster that hosts the service to delete",
+		"deployment-name": "The name of the service to delete",
+	},
 	"stopinstance": {
 		"id": "One or more instance IDs",
 	},
-	"updatebucket":       {},
+	"updatebucket": {},
+	"updatecontainerservice": {
+		"cluster":         "The short name or full Amazon Resource Name (ARN) of the cluster that your service is running on",
+		"deployment-name": "The name of the service to update",
+		"desired-count":   "The number of instantiations of the task to place and keep running in your service",
+		"name":            "The family and revision (family:revision) or full Amazon Resource Name (ARN) of the task definition to run in your service",
+	},
 	"updatedistribution": {},
 	"updateinstance": {
 		"id":   "The ID of the instance",